+++ llvm/lib/IR/AutoUpgrade.cpp
@@ -5956,6 +5956,12 @@ void llvm::UpgradeFunctionAttributes(Function &F) {
     F.removeFnAttr("implicit-section-name");
   }
 
+  if (Attribute A = F.getFnAttribute("nooutline");
+      A.isValid() && A.isStringAttribute()) {
+    F.removeFnAttr("nooutline");
+    F.addFnAttr(Attribute::NoOutline);
+  }
+
   if (!F.empty()) {
     // For some reason this is called twice, and the first time is before any
     // instructions are loaded into the body.